Fix the overflow problem of int type
The overflow occurs when the result of an addition operation
exceeds the maximum value of int type. For example, if the values of
two integers given as inputs are given as 2,147,483,647, the maximum
value of type int, the expected result is 4,294,967,294.

- Support the addition operation is calculated
  normally even if it exceeds the range of int type.

- Add test cases in which the result of addition exceeds
  the maximum value of int type.

Resolves: #1

+++ add.c
@@ -1,4 +1,4 @@
 
-int add(int a, int b){
-	return a + b;
+long long add(int a, int b){
+	return (long long) a + b;
 }
